refactor: chess 패키지 생성
