+++ src/test/java/softeer2nd/chess/PawnTest.java
@@ -1,9 +1,10 @@
-package softeer2nd;
+package softeer2nd.chess;
 
-import org.junit.jupiter.api.*;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 
-import static org.assertj.core.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.*;
 
 public class PawnTest {
     @Test
